rename submodules
