+++ src/sygnet/models.py
@@ -1,4 +1,4 @@
-from .sygnet_requirements import *
+from .requirements import *
 
 logger = logging.getLogger(__name__)
