+++ src/sygnet/train.py
@@ -1,5 +1,5 @@
 ### TRAINING FUNCTIONS
-from .sygnet_requirements import *
+from .requirements import *
 
 logger = logging.getLogger(__name__)
